# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sirene Automatique API Documentation",
        "description": "API Documentation for the Sirene Automatique project",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/sirenes": {
            "get": {
                "tags": [
                    "Sirenes"
                ],
                "summary": "List all sirenes",
                "description": "Lister toutes les sirènes",
                "operationId": "d6660ba4a14cc54751b48746ec218da0",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of sirenes per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Sirene"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Sirenes"
                ],
                "summary": "Create a new sirene (Admin only)",
                "description": "Créer une nouvelle sirène (Admin seulement - génération à l'usine)",
                "operationId": "6ae8c0589d294e9e1ad1d0accf9b111d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateSireneRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sirene created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Sirene"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "This action is unauthorized."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/sirenes/{id}": {
            "get": {
                "tags": [
                    "Sirenes"
                ],
                "summary": "Get sirene details by ID",
                "description": "Afficher les détails d'une sirène",
                "operationId": "ec68e353ca3cf0348cfb99751cd7717b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the sirene to retrieve",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Sirene"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Sirene not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sirene not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Sirenes"
                ],
                "summary": "Update an existing sirene",
                "description": "Mettre à jour une sirène",
                "operationId": "bf1e8cedaa14171fd67923d4afc583f8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the sirene to update",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateSireneRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sirene updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Sirene"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "This action is unauthorized."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Sirene not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sirene not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Sirenes"
                ],
                "summary": "Delete a sirene by ID",
                "description": "Supprimer une sirène",
                "operationId": "3f75f52f33a643d3e841eda4c538558c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the sirene to delete",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Sirene deleted successfully"
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Sirene not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sirene not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/sirenes/numero-serie/{numeroSerie}": {
            "get": {
                "tags": [
                    "Sirenes"
                ],
                "summary": "Get sirene details by serial number",
                "description": "Rechercher une sirène par numéro de série",
                "operationId": "1f0613a79c653f88d47a0f039c881f3c",
                "parameters": [
                    {
                        "name": "numeroSerie",
                        "in": "path",
                        "description": "Serial number of the sirene to retrieve",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Sirene"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Sirene not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sirene not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
        "/api/sirenes/disponibles": {
            "get": {
                "tags": [
                    "Sirenes"
                ],
                "summary": "Get available sirenes (not assigned to any site)",
                "description": "Obtenir les sirènes disponibles (non affectées)",
                "operationId": "dbd50041358d3d767316825fea69231d",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Sirene"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "CreateSireneRequest": {
                "title": "Create Sirene Request",
                "description": "Request body for creating a new sirene",
                "required": [
                    "modele_id",
                    "date_fabrication"
                ],
                "properties": {
                    "modele_id": {
                        "description": "ID of the sirene model",
                        "type": "string",
                        "format": "uuid"
                    },
                    "date_fabrication": {
                        "description": "Manufacturing date of the sirene",
                        "type": "string",
                        "format": "date"
                    },
                    "notes": {
                        "description": "Additional notes for the sirene",
                        "type": "string",
                        "maxLength": 1000,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}